+++ Source/OxyPlot/Annotations/PathAnnotation.cs
@@ -103,28 +103,28 @@ namespace OxyPlot.Annotations
         public double TextPadding { get; set; }
 
         /// <summary>
-        /// Gets or sets the padding of the text border rectangle.
+        /// Gets or sets the padding of the border rectangle.
         /// </summary>
-        /// <value>The text border padding.</value>
-        public OxyThickness TextBorderPadding { get; set; }
+        /// <value>The border padding.</value>
+        public OxyThickness BorderPadding { get; set; }
 
         /// <summary>
-        /// Gets or sets the fill color of the text border rectangle.
+        /// Gets or sets the fill color of the border rectangle.
         /// </summary>
-        /// <value>The text border background color.</value>
-        public OxyColor TextBorderBackground { get; set; }
+        /// <value>The border background color.</value>
+        public OxyColor BorderBackground { get; set; }
 
         /// <summary>
-        /// Gets or sets the stroke color of the text border rectangle.
+        /// Gets or sets the stroke color of the border rectangle.
         /// </summary>
-        /// <value>The text border stroke color.</value>
-        public OxyColor TextBorderStroke { get; set; }
+        /// <value>The border stroke color.</value>
+        public OxyColor BorderStroke { get; set; }
 
         /// <summary>
-        /// Gets or sets the stroke thickness of the text border rectangle.
+        /// Gets or sets the stroke thickness of the border rectangle.
         /// </summary>
-        /// <value>The text border stroke thickness.</value>
-        public double TextBorderStrokeThickness { get; set; }
+        /// <value>The border stroke thickness.</value>
+        public double BorderStrokeThickness { get; set; }
 
         /// <summary>
         /// Gets or sets the text orientation.
@@ -280,16 +280,16 @@ namespace OxyPlot.Annotations
 
                     var textSize = rc.MeasureText(this.Text, this.ActualFont, this.ActualFontSize, this.ActualFontWeight);
 
-                    var textBounds = TextAnnotation.GetTextBounds(position, textSize, this.TextBackgroundPadding, angle, ha, va);
+                    var textBounds = TextAnnotation.GetTextBounds(position, textSize, this.BorderPadding, angle, ha, va);
 
                     if ((angle % 90).Equals(0))
                     {
                         var actualRect = new OxyRect(textBounds[0], textBounds[2]);
-                        rc.DrawRectangle(actualRect, this.TextBackground, this.TextBorderStroke, this.TextBackgroundStrokeThickness, this.EdgeRenderingMode);
+                        rc.DrawRectangle(actualRect, this.BorderBackground, this.BorderStroke, this.BorderStrokeThickness, this.EdgeRenderingMode);
                     }
                     else
                     {
-                        rc.DrawPolygon(textBounds, this.TextBackground, this.TextBorderStroke, this.TextBackgroundStrokeThickness, this.EdgeRenderingMode);
+                        rc.DrawPolygon(textBounds, this.BorderBackground, this.BorderStroke, this.BorderStrokeThickness, this.EdgeRenderingMode);
                     }
 
                     rc.DrawText(